+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -79,6 +79,11 @@ New checks
   Finds potentially incorrect calls to ``memcmp()`` based on properties of the
   arguments.
 
+- New :doc:`cert-smart-pointer-deleter
+  <clang-tidy/checks/cert-smart-pointer-deleter>` check.
+
+  FIXME: add release notes.
+
 - New :doc:`cppcoreguidelines-virtual-class-destructor
   <clang-tidy/checks/cppcoreguidelines-virtual-class-destructor>` check.